+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 -   Allow array type declaration without a primitive type
 -   Added more diagnostic codes to lpc-config.
 -   Allow variable declarations in switch body before first case.
+-   Display current driver type in VSCode status bar
 -   Fix: Parsing error when including a bracketed path with a forward slash
 -   Fix: Struct member declarations can have multiple variables separated by commas
 -   Fix: Multiline defines not tokenized correctly when lines end with `\r\n`